feat(icon-button): convert to typescript
