+++ src/components/icon-button/icon-button.style.ts
@@ -1,10 +1,11 @@
 import styled, { css } from "styled-components";
 import { margin } from "styled-system";
+
 import StyledIcon from "../icon/icon.style";
 import { baseTheme } from "../../style/themes";
 
 const StyledIconButton = styled.button.attrs({ type: "button" })`
-  ${({ disabled }) => css`
+  ${({ disabled }: { disabled?: boolean }) => css`
     ${margin}
     background: transparent;
     border: none;
