--- src/components/icon-button/index.d.ts
@@ -1,2 +1,0 @@
-export { default } from "./icon-button";
-export * from "./icon-button";
