--- src/components/icon-button/index.js
@@ -1,1 +1,0 @@
-export { default } from "./icon-button.component";
