+++ src/components/icon-button/index.ts
@@ -1,0 +1,2 @@
+export { default } from "./icon-button.component";
+export type { IconButtonProps } from "./icon-button.component";
